General typecast. Compiler version independent.

+++ src/MadGraphSrc/read_slha.cc
@@ -81,7 +81,7 @@ void SLHAReader::read_slha_file(string file_name)
 	while (line[0] == ' ')
 	  line = line.substr(1);
 	// Now find end of block name
-	unsigned long space_pos = line.find(' ');
+	size_t space_pos = line.find(' ');
 	if(space_pos != line.npos)
 	  line = line.substr(0, space_pos);
 	block = line;
